requirements.txt
flask
flask-sqlalchemy
pillow
pytest